requirements.txt
streamlit
langchain
langchain-community
PyPDF2
python-docx
pandas
openpyxl
Pillow
torch
diffusers
transformers



